requirements.txt
# Core
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-axes>=6.4
drf-spectacular>=0.27
python-dotenv>=1.0

# Auth
google-auth>=2.28

# Deploy
whitenoise>=6.7
dj-database-url>=2.1
psycopg2-binary>=2.9
gunicorn>=22.0

# Images
Pillow>=10.3
cloudinary>=1.36
django-cloudinary-storage>=0.3

# Search & Filter
django-filter>=24.2

# Payments
razorpay>=1.4

# Testing
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
coverage>=7.5
